% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_color_to_shape.R
\name{add_color_to_shape}
\alias{add_color_to_shape}
\title{Returns a plot of a random shape}
\usage{
add_color_to_shape(shape, shape_color, background_color)
}
\arguments{
\item{shape}{the plot object of the randomly generated shape}

\item{shape_color}{the fill color of the shape}

\item{background_color}{the color of the background of the plot}
}
\value{
a modified plot of the shape
}
\description{
Returns a plot of a random shape
}
