% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{add_text}
\alias{add_text}
\title{Returns a string annotation}
\usage{
add_text(plot, text, size)
}
\arguments{
\item{plot}{the plot}

\item{text}{string of text}

\item{size}{the size of the font}
}
\value{
a scatterplot of circles
}
\description{
Returns a string annotation
}
