% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confCircles.R
\name{confetti_circle}
\alias{confetti_circle}
\title{Returns a plot of randomly generated circle(s)}
\usage{
confetti_circle(plot, n, color, size)
}
\arguments{
\item{plot}{the plot to which to apply the circles}

\item{n}{the number of circles}

\item{color}{the color of the circles}

\item{size}{the size of the circles}
}
\value{
a scatterplot of circles
}
\description{
Returns a plot of randomly generated circle(s)
}
