% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confDiamonds.R
\name{confetti_diamond}
\alias{confetti_diamond}
\title{Returns a plot of randomly generated diamond(s)}
\usage{
confetti_diamond(plot, n, color, size)
}
\arguments{
\item{plot}{the plot to which to apply the diamonds}

\item{n}{the number of diamonds}

\item{color}{the color of the diamonds}

\item{size}{the size of the diamonds}
}
\value{
a scatterplot of diamonds
}
\description{
Returns a plot of randomly generated diamond(s)
}
