% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confStar.R
\name{confetti_star}
\alias{confetti_star}
\title{Returns a plot of randomly generated star(s)}
\usage{
confetti_star(plot, n, color, size)
}
\arguments{
\item{plot}{the plot to which to apply the stars}

\item{n}{the number of stars}

\item{color}{the color of the stars}

\item{size}{the size of the stars}
}
\value{
a scatterplot of stars
}
\description{
Returns a plot of randomly generated star(s)
}
