% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_shape.R
\name{generate_random_shape}
\alias{generate_random_shape}
\title{Returns a plot of a random shape}
\usage{
generate_random_shape(num_points, num_cycles)
}
\arguments{
\item{num_points}{the the number of points in the shape}

\item{num_cycles}{the number of cycles that shape has}
}
\value{
a plot of the random shape
}
\description{
Returns a plot of a random shape
}
